+++ pytest_pyodide/fixture.py
@@ -10,6 +10,7 @@ from .runner import (
     PlaywrightFirefoxRunner,
     SeleniumChromeRunner,
     SeleniumFirefoxRunner,
+    SeleniumSafariRunner,
     _BrowserBaseRunner,
 )
 from .server import spawn_web_server
@@ -75,6 +76,7 @@ def selenium_common(
     browser_set = {
         ("selenium", "firefox"): SeleniumFirefoxRunner,
         ("selenium", "chrome"): SeleniumChromeRunner,
+        ("selenium", "safari"): SeleniumSafariRunner,
         ("selenium", "node"): NodeRunner,
         ("playwright", "firefox"): PlaywrightFirefoxRunner,
         ("playwright", "chrome"): PlaywrightChromeRunner,
@@ -113,6 +115,30 @@ def selenium_standalone(request, runtime, web_server_main, playwright_browsers):
                 yield selenium
             finally:
                 print(selenium.logs)
+
+
+@pytest.fixture(scope="function")
+def selenium_standalone_refresh(selenium):
+    """
+    Experimental standalone fixture which refreshes a page instead of
+    instantiating a new webdriver session.
+
+    by setting STANDALONE_REFRESH env variable,
+    selenium_standalone_refresh fixture will override selenium_standalone
+    """
+    selenium.clean_logs()
+
+    yield selenium
+
+    selenium.refresh()
+    selenium.load_pyodide()
+    selenium.initialize_pyodide()
+    selenium.save_state()
+    selenium.restore_state()
+
+
+if os.environ.get("STANDALONE_REFRESH"):
+    selenium_standalone = selenium_standalone_refresh  # type: ignore[assignment] # noqa: F811
 
 
 @pytest.fixture(scope="module")