+++ pytest_pyodide/runner.py
@@ -100,6 +100,7 @@ class JavascriptException(Exception):
 
 class _BrowserBaseRunner:
     browser = ""
+    script_timeout = 20
     JavascriptException = JavascriptException
 
     def __init__(
@@ -108,7 +109,6 @@ class _BrowserBaseRunner:
         server_hostname="127.0.0.1",
         server_log=None,
         load_pyodide=True,
-        script_timeout=20,
         script_type="classic",
         dist_dir=None,
         *args,
@@ -121,8 +121,7 @@ class _BrowserBaseRunner:
         self.script_type = script_type
         self.dist_dir = dist_dir
         self.driver = self.get_driver()
-        self.set_script_timeout(script_timeout)
-        self.script_timeout = script_timeout
+        self.set_script_timeout(self.script_timeout)
         self.prepare_driver()
         self.javascript_setup()
         if load_pyodide:
@@ -321,6 +320,7 @@ class _SeleniumBaseRunner(_BrowserBaseRunner):
 
     def set_script_timeout(self, timeout):
         self.driver.set_script_timeout(timeout)
+        self.script_timeout = timeout
 
     def quit(self):
         self.driver.quit()
@@ -433,6 +433,19 @@ class SeleniumChromeRunner(_SeleniumBaseRunner):
         self.driver.execute_cdp_cmd("HeapProfiler.collectGarbage", {})
 
 
+class SeleniumSafariRunner(_SeleniumBaseRunner):
+
+    browser = "safari"
+    script_timeout = 30
+
+    def get_driver(self):
+        from selenium.webdriver import Safari
+        from selenium.webdriver.safari.options import Options
+
+        options = Options()
+        return Safari(options=options)
+
+
 class PlaywrightChromeRunner(_PlaywrightBaseRunner):
     browser = "chrome"
 
@@ -487,7 +500,7 @@ class NodeRunner(_BrowserBaseRunner):
         pass
 
     def set_script_timeout(self, timeout):
-        self._timeout = timeout
+        self.script_timeout = timeout
 
     def quit(self):
         self.p.sendeof()
@@ -523,7 +536,7 @@ class NodeRunner(_BrowserBaseRunner):
         self.p.sendline(cmd_id)
         self.p.sendline(wrapped)
         self.p.sendline(cmd_id)
-        self.p.expect_exact(f"{cmd_id}:UUID\r\n", timeout=self._timeout)
+        self.p.expect_exact(f"{cmd_id}:UUID\r\n", timeout=self.script_timeout)
         self.p.expect_exact(f"{cmd_id}:UUID\r\n")
         if self.p.before:
             self._logs.append(self.p.before.decode()[:-2].replace("\r", ""))