+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.22.0] - 2022.09.02
+
+- Add `selenium_standalone_refresh` fixture ([#27](https://github.com/pyodide/pytest-pyodide/pull/27))
+- Add selenium safari support ([#27](https://github.com/pyodide/pytest-pyodide/pull/27))
+
 ## [0.21.1] - 2022.08.10
 
 - Use forked tblib and clean up tracebacks ([#22](https://github.com/pyodide/pytest-pyodide/pull/22))